feat: реализована работа с режиссерами в FilmDbStorage
- Добавлена поддержка сохранения режиссеров в методах addFilm и renewFilm.
- Внедрена очистка и обновление связей в таблице film_directors при обновлении фильма.
- Добавлен метод getFilmsByDirector с поддержкой сортировки по годам и лайкам.
- Проведен рефакторинг: удален избыточный маппинг данных в FilmDbStorage в пользу переиспользования метода returnFilmByID.

+++ src/main/resources/schema.sql
@@ -49,7 +49,7 @@ CREATE TABLE IF NOT EXISTS directors (
     );
 
 CREATE TABLE IF NOT EXISTS film_directors (
-    film_id BIGINT REFERENCES films(film_id) ON DELETE CASCADE,
-    director_id INT REFERENCES directors(director_id) ON DELETE CASCADE,
+    film_id INTEGER REFERENCES films(film_id) ON DELETE CASCADE,
+    director_id INTEGER REFERENCES directors(director_id) ON DELETE CASCADE,
     PRIMARY KEY (film_id, director_id)
     );
